Don't always prompt for favorite and sort favorites at startup

+++ pytoggl.py
@@ -39,20 +39,21 @@ headers = {
     "Content-Type": "application/json",
     "Authorization": "Basic %s" % auth.decode("utf-8"),
 }
-
+USED_FAVORITE = False
 chosen_fav = "-1"
 if FAV_PROJECTS_FILE.is_file():
     print("Favorites:\n")
     print(30 * "-")
     with open(FAV_PROJECTS_FILE, "r") as fav_projects_file:
-        fav_projects = fav_projects_file.readlines()
+        fav_projects = sorted(fav_projects_file.readlines())
         for idx, fav in enumerate(fav_projects):
-            print("[%s] %s" % (idx, fav.split("\t")[1]))
+            print("[%s] %s" % (idx, fav.strip().split("\t")[1]))
         while chosen_fav not in [str(i) for i in range(len(fav_projects))] + [""]:
             chosen_fav = input(
                 "Choose favorite project (leave empty for manual setup): "
             )
             if chosen_fav != "":
+                USED_FAVORITE = True
                 sel_project = fav_projects[int(chosen_fav)]
                 sel_project_id = sel_project.split("\t")[0]
                 sel_project_name = sel_project.split("\t")[1]
@@ -157,22 +158,23 @@ if submit_confirmation.lower() == "y":
     if "data" in time_entry_response.keys():
         print("Time entry submitted succesfully.")
 
-    project_confirmation = ""
-    while project_confirmation.lower() not in ("y", "n"):
-        project_confirmation = input("Save project to favorites? [y/n]: ")
+    if not USED_FAVORITE:
+        project_confirmation = ""
+        while project_confirmation.lower() not in ("y", "n"):
+            project_confirmation = input("Save project to favorites? [y/n]: ")
 
-    if project_confirmation.lower() == "y":
-        if FAV_PROJECTS_FILE.is_file():
-            with open(FAV_PROJECTS_FILE, "r") as fav_projects_file:
-                fav_projects = set(fav_projects_file.readlines())
-                fav_projects.add("%s\t%s" % (sel_project_id, sel_project_name))
-        else:
-            fav_projects = {"%s\t%s" % (sel_project_id, sel_project_name)}
-        with open(FAV_PROJECTS_FILE, "w+") as fav_projects_file:
-            for fav in list(fav_projects)[:-1]:
-                fav_projects_file.write("%s\n" % fav.strip())
-            fav_projects_file.write("%s" % list(fav_projects)[-1].strip())
-        print("Project saved to favorites.")
+        if project_confirmation.lower() == "y":
+            if FAV_PROJECTS_FILE.is_file():
+                with open(FAV_PROJECTS_FILE, "r") as fav_projects_file:
+                    fav_projects = set(fav_projects_file.readlines())
+                    fav_projects.add("%s\t%s" % (sel_project_id, sel_project_name))
+            else:
+                fav_projects = {"%s\t%s" % (sel_project_id, sel_project_name)}
+            with open(FAV_PROJECTS_FILE, "w+") as fav_projects_file:
+                for fav in list(fav_projects)[:-1]:
+                    fav_projects_file.write("%s\n" % fav.strip())
+                fav_projects_file.write("%s" % list(fav_projects)[-1].strip())
+            print("Project saved to favorites.")
 
 
 else:
